% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetExperiment-ampliPanel.R
\docType{data}
\name{ampliPanel}
\alias{ampliPanel}
\title{An amplicon panel example for use the TarSeqQC R package.}
\format{A TargetExperiment object}
\source{
see \code{\link{TargetExperiment-class}}
}
\description{
A non-real dataset containing amplicon sequencing results to test the TarSeqQC
package.
}
\details{
\describe{
\item{bedFile}{Bed file containing 29 amplicons and 8 genes.} 
\item{feature}{Character "amplicon" indicating that the analyzed features 
are amplicon sequences} 
\item{attribute}{Character "coverage"}
}
}
\author{
Gabriela A. Merino \email{gmerino@bdmg.com.ar}, Cristobal Fresno
\email{cfresno@bdmg.com.ar}, and Elmer A. Fernandez 
\email{efernandez@bdmg.com.ar}
}
\seealso{
Other TargetExperiment: \code{\link{TargetExperiment-class}},
    \code{\link{TargetExperiment}},
    \code{\link{ampliPanel2}}, \code{\link{initialize}},
    \code{\link{myCounts}}
}

