% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetExperiment-pileupCounts.R
\docType{methods}
\name{checkBedFasta}
\alias{checkBedFasta}
\alias{checkBedFasta-methods}
\title{Function to control Bed and FASTA files compatibility.}
\usage{
checkBedFasta(bedFile, fastaFile)
}
\arguments{
\item{bedFile}{Character indicating the bed file full path.}

\item{fastaFile}{Character indicating the full path to the genome reference 
    file.}
}
\value{
NULL}
\description{
\code{checkBedFasta} checks the compatibility of a Bed file and a Fasta file.
The functions first will control the consistency of the Bed file in terms of
duplicated positions or feature's IDs and correct definition of start-end
values. Then, the method will control the consistency between the specified
features and the reference file. During its execution, several testing 
messages will be printed.
}
\examples{
##Define the bed and fasta file full paths
bedFile<-system.file("extdata", "mybed.bed", package="TarSeqQC", mustWork=TRUE)
fastaFile<-system.file("extdata", "myfasta.fa", package="TarSeqQC",
    mustWork=TRUE)
##Checking the bed-fasta consistency
checkBedFasta(bedFile, fastaFile)
}
\author{
Gabriela A. Merino \email{gmerino@bdmg.com.ar}, Cristobal Fresno
\email{cfresno@bdmg.com.ar}, Yanina Murua \email{ymurua@leloir.org.ar},
Andrea S. Llera \email{allera@leloir.org.ar} and Elmer A. Fernandez 
\email{efernandez@bdmg.com.ar}
}
\seealso{
\code{\link{TargetExperiment-class}}
}

