% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetExperiment-myCounts.R
\docType{data}
\name{myCounts}
\alias{myCounts}
\title{A pileup matrix example for use the TarSeqQC R package.}
\format{A data.frame object}
\source{
see \code{\link{TargetExperiment-class}}
}
\description{
The pileup matrix obtained using pileupCounts. It is built on the non-real
dataset containing amplicon sequencing results to test the TarSeqQC package.
}
\details{
\describe{
\item{pos}{genomic positions of the explored features.} 
\item{seqnames}{chromosomes of the explored features.} 
\item{seq}{reference nucleotide corresponding to the genomic position.}
\item{A,C,G,T,N}{number of nucleotide read.}
\item{=}{Amount of read nucleotides matching the reference nucleotide.}
\item{-}{Amount of read deletions.}
\item{which_label}{feature location.}
\item{counts}{Total read counts}
}
}
\author{
Gabriela A. Merino \email{gmerino@bdmg.com.ar}, Cristobal Fresno
\email{cfresno@bdmg.com.ar}, Yanina Murua \email{ymurua@leloir.org.ar},
Andrea S. Llera \email{allera@leloir.org.ar} and Elmer A. Fernandez 
\email{efernandez@bdmg.com.ar}
}
\seealso{
Other TargetExperiment: \code{\link{TargetExperiment-class}},
    \code{\link{TargetExperiment}},
    \code{\link{ampliPanel2}}, \code{\link{ampliPanel}},
    \code{\link{initialize}}
}

