% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetExperiment-plotInOutFeatures.R
\docType{methods}
\name{plotInOutFeatures}
\alias{plotInOutFeatures}
\alias{plotInOutFeatures,TargetExperiment-method}
\alias{plotInOutFeatures,data.frame-method}
\title{Function to explore read percentages in targeted regions and out 
targeted regions.}
\usage{
plotInOutFeatures(object, ...)

\S4method{plotInOutFeatures}{data.frame}(object, absolute = FALSE)

\S4method{plotInOutFeatures}{TargetExperiment}(object, absolute = FALSE,
    BPPARAM = bpparam())
}
\arguments{
\item{object}{a data frame or a TargetExperiment.}

\item{...}{additional parameters according to the function call}

\item{absolute}{logical indicating if absolute frequency should be used.}

\item{BPPARAM}{An optional BiocParallelParam instance defining the parallel
back-end to be used during evaluation.}
}
\value{
ggplot object.
}
\description{
\code{plotInOutFeatures} allows the graphical exploration of the data 
frame obtained using \code{readFrequencies}. This data frame contains
information about the amount of reads mapped to the targeted regions and out of
them. This information is presented in rows, one for each chromosome and in 
absolute and relative amounts. After it invocation, a bar plot built as a 
ggplot object is returned
}
\note{
see full example in \code{\link{TargetExperiment-class}}
}
\examples{
## loading TargetExperiment object
data(ampliPanel, package="TarSeqQC")
## Defining bam file, bed file and fasta file names and paths
setBamFile(ampliPanel)<-system.file("extdata", "mybam.bam", 
    package="TarSeqQC", mustWork=TRUE)
setFastaFile(ampliPanel)<-system.file("extdata", "myfasta.fa", 
    package="TarSeqQC", mustWork=TRUE)

g<-plotInOutFeatures(ampliPanel)

}
\author{
Gabriela A. Merino \email{gmerino@bdmg.com.ar}, Cristobal Fresno
\email{cfresno@bdmg.com.ar}, Yanina Murua \email{ymurua@leloir.org.ar},
Andrea S. Llera \email{allera@leloir.org.ar} and Elmer A. Fernandez 
\email{efernandez@bdmg.com.ar}
}

