% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetExperiment-biasExploration.R
\docType{methods}
\name{biasExploration}
\alias{biasExploration}
\alias{biasExploration,TargetExperiment-method}
\alias{biasExploration-methods}
\title{Plot attribute density and boxplot for each bias source quartile or 
category.}
\usage{
biasExploration(object, source = c("length", "gc", "pool"), dens = FALSE)

\S4method{biasExploration}{TargetExperiment}(object, source = c("length",
    "gc", "pool"), dens = FALSE)
}
\arguments{
\item{object}{TargetExperiment class object.}

\item{source}{Character 'gc','length', or 'pool' indicating the source bias. 
In the case of 'gc' and 'length', it will be categorized in four groups 
according to its quartiles. In the case of 'pool', its groups will be
conserved.}

\item{dens}{Logical indicating if density plot should be added using the 
geom_violin ggplot2 method.}
}
\value{
ggplot2 graphics.
}
\description{
\code{biasExploration} plots density and box-plot of the analyzed attribute 
for each bias source' quartiles per categories. It helps the identification 
of some bias due to high source values, for example, high gc content. This 
graphics could plot together using the ggplot2 geom_violin method.
}
\note{
see full example in \code{\link{TargetExperiment-class}}
}
\examples{
## Loading the TargetExperiment object
data(ampliPanel, package="TarSeqQC")

# Attribute boxplot and density plot exploration
g<-biasExploration(ampliPanel,source="gc", dens=TRUE)
# x11(type="cairo")
if(interactive()){
g
}
}
\author{
Gabriela A. Merino \email{gmerino@bdmg.com.ar}, Cristobal Fresno
\email{cfresno@bdmg.com.ar}, Yanina Murua \email{ymurua@leloir.org.ar},
Andrea S. Llera \email{allera@leloir.org.ar} and Elmer A. Fernandez 
\email{efernandez@bdmg.com.ar}
}
\seealso{
\code{\link{plot}}, \code{\link{plotFeatPerform}}
}

