% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TargetExperiment-addStatSummSheet.R,
% R/TargetExperiment-buildReport.R, R/TargetExperiment-ggplotColours.R
\docType{methods}
\name{addStatSummSheet}
\alias{addStatSummSheet}
\alias{addStatSummSheet,TargetExperiment-method}
\alias{addStatSummSheet-methods}
\alias{buildReport}
\alias{buildReport,TargetExperiment-method}
\alias{buildReport-methods}
\alias{ggplotColours}
\alias{ggplotColours,TargetExperiment-method}
\alias{ggplotColours-methods}
\title{Build excel report of the Target Experiment.}
\usage{
addStatSummSheet(object, wb, attributeThres = c(0, 1, 50, 200, 500, Inf),
    imageFile)

\S4method{addStatSummSheet}{TargetExperiment}(object, wb,
    attributeThres = c(0, 1, 50, 200, 500, Inf), imageFile)

buildReport(object, attributeThres = c(0, 1, 50, 200, 500, Inf),
    imageFile = NULL, file)

\S4method{buildReport}{TargetExperiment}(object, attributeThres = c(0, 1, 50,
    200, 500, Inf), imageFile = NULL, file = "Results.xlsx")

ggplotColours(object, n)

\S4method{ggplotColours}{TargetExperiment}(object, n)
}
\arguments{
\item{object}{TargetExperiment class object.}

\item{wb}{A workbook object that will contain the report.}

\item{attributeThres}{Numeric indicating the intervals extreme values.}

\item{imageFile}{Character indicating the name of the file that contains
the plot that could be insert in the report.}

\item{file}{Character indicating the name of the report.}

\item{n}{amount of colors.}
}
\value{
Workbook object.

NULL.

colours
}
\description{
\code{addStatSummSheet} adds the statistics summary sheet to the workbook
that contains the Target Experiment Report.



\code{buildReport} builds an excel file containing some statistical results.
These are computed to the selected attribute (e.g. 'coverage') along features
(e.g. 'amplicon') and genes. If 'imageFile' is null, the graph generated
calling the generic plot function will be used.



\code{ggplotColours} is a function to know what color is used when
ggplot is called.


}
\note{
see full example in \code{\link{TargetExperiment-class}}
}
\examples{
## Loading the TargetExperiment object
data(ampliPanel,package="TarSeqQC")
# definition of the interval extreme values
attributeThres<-c(0,1,50,200,500, Inf)

## Building the XLSX report
imageFile<-system.file("extdata", "plot.pdf", package="TarSeqQC",
mustWork=TRUE)
buildReport(ampliPanel, attributeThres=attributeThres, imageFile=imageFile,
file="results.xlsx")
}
\author{
Gabriela A. Merino \email{gmerino@bdmg.com.ar}, Cristobal Fresno
\email{cfresno@bdmg.com.ar} and Elmer A. Fernandez
\email{efernandez@bdmg.com.ar}
}
\seealso{
\code{\link{TargetExperiment-class}}
}

