% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetExperiment-constructor.R
\docType{methods}
\name{TargetExperiment}
\alias{TargetExperiment}
\alias{TargetExperiment-methods}
\title{TargetExperiment constructor}
\usage{
TargetExperiment(bedFile, bamFile, fastaFile, scanBamP = NULL,
    pileupP = NULL, feature = NULL, attribute = NULL, BPPARAM = bpparam())
}
\arguments{
\item{bedFile}{Character indicating the bed file full path.}

\item{bamFile}{Character indicating the alignment and index bam
    files full path.}

\item{fastaFile}{Character indicating the full path to the genome reference
    and index files.}

\item{scanBamP}{ScanBamParam indicating the parameters the BAM file read.}

\item{pileupP}{PileupParam indicating the parameters for the pileup build.}

\item{feature}{Character indicating the name of the feature that will
    be explored (e.g 'amplicon', 'exon').}

\item{attribute}{Character indicating the name of the attribute that will
    be explored. Should be 'coverage' or 'medianCounts'.}

\item{BPPARAM}{An optional BiocParallelParam instance defining the parallel
    back-end to be used during evaluation.}
}
\value{
TargetExperiment object.
}
\description{
\code{TargetExperiment} creates a TargetExperiment object with the
architecture specified by the bed and alignment BAM files. If 'scanBamP' 
and/or 'pileupP' parameters are not specified, default values of their 
constructors will be used. attribute and feature parameters can be set 
after constructor calling.
}
\note{
see full example in \code{\link{TargetExperiment-class}}
}
\examples{
## Defining bam file, bed file and fasta file names and paths
bamFile<-system.file("extdata", "mybam.bam", package="TarSeqQC",
    mustWork=TRUE)
bedFile<-system.file("extdata", "mybed.bed", package="TarSeqQC",
    mustWork=TRUE)
fastaFile<-system.file("extdata", "myfasta.fa", package="TarSeqQC", 
    mustWork=TRUE)

## Creating a TargetExperiment object

# Defining feature parameter
feature<-"amplicon"
# Defining attribute parameter
attribute<-"coverage"
##Calling the constructor
object<-TargetExperiment(bedFile, bamFile, fastaFile, attribute=attribute,
feature=feature)

}
\author{
Gabriela A. Merino \email{gmerino@bdmg.com.ar}, Cristobal Fresno
\email{cfresno@bdmg.com.ar}, Yanina Murua \email{ymurua@leloir.org.ar},
Andrea S. Llera \email{allera@leloir.org.ar} and Elmer A. Fernandez 
\email{efernandez@bdmg.com.ar}
}
\seealso{
\code{\link{TargetExperiment-class}}1

Other TargetExperiment: \code{\link{TargetExperiment-class}},
    \code{\link{ampliPanel2}}, \code{\link{ampliPanel}},
    \code{\link{initialize}}, \code{\link{myCounts}}
}

