% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TargetExperiment-initialize.R
\docType{methods}
\name{initialize}
\alias{initialize}
\alias{initialize,TargetExperiment-method}
\title{TargetExperiment object constructor.}
\usage{
\S4method{initialize}{TargetExperiment}(.Object, bedFile, bamFile, fastaFile,
    scanBamP = NULL, pileupP = NULL, feature = NULL, attribute = NULL,
    BPPARAM = bpparam())
}
\arguments{
\item{.Object}{TargetExperiment class.}

\item{bedFile}{Character indicating the bed file full path.}

\item{bamFile}{Character indicating the alignment and index bam
files full paths.}

\item{fastaFile}{Character indicating the full path to the genome reference
and index files.}

\item{scanBamP}{ScanBamParam indicating the parameters for read the BAM
file.}

\item{pileupP}{PileupParam indicating the parameters for pileup building.}

\item{feature}{Character indicating the name of the feature that will
be explored (e.g 'amplicon', 'exon').}

\item{attribute}{Character indicating the name of the attribute that will
be explored. Should be 'coverage' or 'medianCounts'.}

\item{BPPARAM}{An optional BiocParallelParam instance defining the parallel
back-end to be used during evaluation.}
}
\value{
TargetExperiment object.
}
\description{
\code{initialize} creates the TargetExperiment object architecture for the
specified bed and alingment BAM files. If 'scanBamP' and/or 'pileupP'
parameters are not specified, default values of their constructors will be
used.
}
\note{
see full example in \code{\link{TargetExperiment-class}}
}
\examples{
## Defining bam file, bed file and fasta file names and paths
if (interactive()){
    bamFile<-system.file("extdata", "mybam.bam", package="TarSeqQC",
        mustWork=TRUE)
    bedFile<-system.file("extdata", "mybed.bed", package="TarSeqQC",
        mustWork=TRUE)
    fastaFile<-system.file("extdata", "myfasta.fa", package="TarSeqQC",
        mustWork=TRUE)

    ## Creating a TargetExperiment object

    ## Defining feature parameter
    feature<-"amplicon"
    ## Defining attribute parameter
    attribute<-"coverage"
    ##Calling the constructor
    ampliPanel<-TargetExperiment(bedFile, bamFile, fastaFile, 
        attribute=attribute, feature=feature)
}
}
\author{
Gabriela A. Merino \email{gmerino@bdmg.com.ar}, Cristobal Fresno
\email{cfresno@bdmg.com.ar} and Elmer A. Fernandez
\email{efernandez@bdmg.com.ar}
}
\seealso{
\code{\link{TargetExperiment}}, \code{\link{buildFeaturePanel}}
\code{\link{summarizePanel}}

Other TargetExperiment: \code{\link{TargetExperiment-class}};
    \code{\link{TargetExperiment}},
    \code{\link{TargetExperiment-methods}};
    \code{\link{ampliPanel}}
}

