% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TargetExperiment-plot.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,TargetExperiment,ANY-method}
\title{Plot TargetExperiment object overview.}
\usage{
\S4method{plot}{TargetExperiment,ANY}(x, y, attributeThres = c(0, 1, 50, 200,
    500, Inf), binSize = 1, spaceGene = 0.2, spaceChr = 1.2,
    innerRadius = 0.3, outerRadius = 1, guides = c(20, 40, 60, 80),
    alphaStart = -0.3, circleProportion = 0.95, direction = "inwards",
    chrLabels = FALSE)
}
\arguments{
\item{x}{TargetExperiment class object.}

\item{y}{not used but necessary for redefining the generic function.}

\item{attributeThres}{Numeric indicating the interval extreme values.}

\item{binSize}{Numeric indicating bin width. Should probably be left
as 1, as other parameters are relative to it.}

\item{spaceGene}{Numeric. Space between bins.}

\item{spaceChr}{Numeric. Space between chromosomes.}

\item{innerRadius}{Numeric. Radius of inner circle.}

\item{outerRadius}{Numeric. Radius of outer circle.}

\item{guides}{A vector with percentages to use for the white guide lines.}

\item{alphaStart}{Numeric offset from 12 o'clock in radians.}

\item{circleProportion}{Numeric proportion of the circle to cover.}

\item{direction}{Character indicating if the increasing count goes from or
to the centre.}

\item{chrLabels}{Logical. Chromosome names must be plotted?.}
}
\value{
a ggplot2 graph.
}
\description{
\code{plot} allows a fast and simple representation of one feature panel
using a polar histogram plot. Histogram bar reflects the percentage of
features that have shown the analyzed attribute in a user setted interval.
The resulting graph can be busy and might be better off saved.
}
\note{
see full example in \code{\link{TargetExperiment-class}}
}
\examples{
if(interactive()){
## Loading the TargetExperiment object
data(ampliPanel, package="TarSeqQC")
# Definition of the interval extreme values
attributeThres<-c(0,1,50,200,500, Inf)

## Plot panel overview
g<-plot(ampliPanel, attributeThres, chrLabels =TRUE)
g
}
}
\author{
Gabriela A. Merino \email{gmerino@bdmg.com.ar} Cristobal Fresno
\email{cfresno@bdmg.com.ar} and Elmer A. Fernandez
\email{efernandez@bdmg.com.ar}
}
\references{
\url{http://www.r-bloggers.com/polar-histogram-pretty-and-useful/}
}
\seealso{
\code{\link{plotFeatPerform}}
}

