% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetExperiment-plotAttrExpl.R, 
% R/TargetExperimentList-plotAttrExpl.R
\docType{methods}
\name{plotAttrExpl}
\alias{plotAttrExpl}
\alias{plotAttrExpl,TargetExperiment-method}
\alias{plotAttrExpl,TargetExperimentList-method}
\alias{plotAttrExpl-methods}
\title{Plot attribute exploration of a TargetExperiment/TargetExperimentList
    object.}
\usage{
plotAttrExpl(object, dens = FALSE, join = FALSE, log = TRUE,
    pool = FALSE, ...)

\S4method{plotAttrExpl}{TargetExperiment}(object, level = "feature",
    join = TRUE, log = TRUE, color = "blue")

\S4method{plotAttrExpl}{TargetExperimentList}(object, dens = FALSE,
    join = FALSE, log = TRUE, pool = FALSE, attributeThres = NULL)
}
\arguments{
\item{object}{TargetExperiment/TargetExperimentList class object.}

\item{dens}{Logical indicating if density plot should be included}

\item{join}{Logical indicating if boxplot and density function should be 
plotted together using the ggplot2 geom_violin method.}

\item{log}{Logical indicating if the attribute should be considered in 
log10 scale.}

\item{pool}{Logical indicating if plots should be displayed for each pool
separately}

\item{...}{necessary arguments}

\item{level}{Character 'feature' or 'gene' indicating at which level should 
be analyzed the attribute.}

\item{color}{A character indicating a valid name color.}

\item{attributeThres}{Numeric indicating the attribute interval extreme 
values. It is not a mandatory parameter but if it is specified,then the 
plots will be colored according to the interval in which falls the attribute
median values.}
}
\value{
ggplot2 graphics.
}
\description{
\code{plotAttrExpl} plots density and/or box-plot of the analyzed attribute 
at a feature level. These graphics could be displayed together using the 
ggplot2 geom_violin method. If panel's pools are present, one facet for each
pool  will be showed.
}
\note{
see full example in \code{\link{TargetExperiment-class}}
}
\examples{
## Loading the TargetExperiment object
data(ampliPanel, package="TarSeqQC")

# Attribute boxplot and density plot exploration
g<-plotAttrExpl(ampliPanel,level="feature",join=TRUE, log=FALSE, color="blue")
# x11(type="cairo")
if(interactive()){
g
}
## Loading the TargetExperimentList object
data(TEList, package="TarSeqQC")

# Attribute boxplot and density plot exploration
g<-plotAttrExpl(TEList, log=FALSE, pool=FALSE)
# x11(type="cairo")
if(interactive()){
g
}

}
\author{
Gabriela A. Merino \email{gmerino@bdmg.com.ar}, Cristobal Fresno
\email{cfresno@bdmg.com.ar}, Yanina Murua \email{ymurua@leloir.org.ar},
Andrea S. Llera \email{allera@leloir.org.ar} and Elmer A. Fernandez 
\email{efernandez@bdmg.com.ar}
}
\seealso{
\code{\link{plot}}, \code{\link{plotFeatPerform}}
}

