% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetExperiment-plotAttrPerform.R
\docType{methods}
\name{plotAttrPerform}
\alias{plotAttrPerform}
\alias{plotAttrPerform,TargetExperiment-method}
\alias{plotAttrPerform-methods}
\title{Plot feature performance of a TargetExperiment object.}
\usage{
plotAttrPerform(object, attributeThres = c(0, 1, 50, 200, 500, Inf))

\S4method{plotAttrPerform}{TargetExperiment}(object, attributeThres = c(0, 1,
    50, 200, 500, Inf))
}
\arguments{
\item{object}{TargetExperiment class object.}

\item{attributeThres}{Numeric indicating the intervals extreme values.}
}
\value{
ggplot2 graphics
}
\description{
\code{plotAttrPerform} plots the achieved performance for the selected 
attribute. The resulting graph shows one bar per each attribute interval and
its height is defined according to the number of features achieving attribute
values within that interval.
}
\note{
see full example in \code{\link{TargetExperiment-class}}
}
\examples{
## Loading the TargetExperiment object
data(ampliPanel, package="TarSeqQC")

# Definition of the interval extreme values
attributeThres<-c(0,1,50,200,500, Inf)

# Plot panel overview in a feature performance plot
g<-plotAttrPerform(ampliPanel, attributeThres=attributeThres)
if(interactive()){
g
}
}
\author{
Gabriela A. Merino \email{gmerino@bdmg.com.ar}, Cristobal Fresno
\email{cfresno@bdmg.com.ar}, Yanina Murua \email{ymurua@leloir.org.ar},
Andrea S. Llera \email{allera@leloir.org.ar} and Elmer A. Fernandez 
\email{efernandez@bdmg.com.ar}
}
\seealso{
\code{\link{plot}}
}

