% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetExperiment-plotFeatPerform.R
\docType{methods}
\name{plotFeatPerform}
\alias{plotFeatPerform}
\alias{plotFeatPerform,TargetExperiment-method}
\alias{plotFeatPerform-methods}
\title{Plot feature performance of a TargetExperiment object.}
\usage{
plotFeatPerform(object, attributeThres = c(0, 1, 50, 200, 500, Inf),
    complete = TRUE, log = TRUE, featureLabs = FALSE, sepChr = FALSE,
    legend = TRUE)

\S4method{plotFeatPerform}{TargetExperiment}(object, attributeThres = c(0, 1,
    50, 200, 500, Inf), complete = TRUE, log = TRUE, featureLabs = FALSE,
    sepChr = FALSE, legend = TRUE)
}
\arguments{
\item{object}{TargetExperiment class object.}

\item{attributeThres}{Numeric indicating the intervals extreme values.}

\item{complete}{Logical indicating if the gene and feature level exploration
should be plotted.}

\item{log}{Logical indicating if the attribute should be considered in 
log10 scale.}

\item{featureLabs}{Logical indicating if feature labels should be plotted.}

\item{sepChr}{Logical indicating if the plot should show chromosome 
divisions.}

\item{legend}{Logical indicating if legend should be plotted.}
}
\value{
ggplot2 graphics
}
\description{
\code{plotFeatPerform} plots the achieved performance for each feature/gene.
The resulting graph shows one bar per each feature/gene with height 
according to its attribute value. If complete is set as TRUE, two bar plots
(feature and gene level) will be stored in the resulting ggplot object.
}
\note{
see full example in \code{\link{TargetExperiment-class}}
}
\examples{
## Loading the TargetExperiment object
data(ampliPanel, package="TarSeqQC")

# Definition of the interval extreme values
attributeThres<-c(0,1,50,200,500, Inf)

# Plot panel overview in a feature performance plot
g<-plotFeatPerform(ampliPanel, attributeThres=attributeThres, log=FALSE, 
featureLabs=TRUE, sepChr=TRUE, legend=TRUE)
if(interactive()){
g
}
}
\author{
Gabriela A. Merino \email{gmerino@bdmg.com.ar}, Cristobal Fresno
\email{cfresno@bdmg.com.ar}, Yanina Murua \email{ymurua@leloir.org.ar},
Andrea S. Llera \email{allera@leloir.org.ar} and Elmer A. Fernandez 
\email{efernandez@bdmg.com.ar}
}
\seealso{
\code{\link{plot}}
}

