% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TargetExperiment-plotFeature.R
\docType{methods}
\name{plotFeature}
\alias{plotFeature}
\alias{plotFeature,TargetExperiment-method}
\alias{plotFeature-methods}
\title{Plot read profiles for a particular feature.}
\usage{
plotFeature(object, featureID, SNPs = TRUE, xlab = "", title = "",
    size = 0.5, BPPARAM = bpparam())

\S4method{plotFeature}{TargetExperiment}(object, featureID, SNPs = TRUE,
    xlab = "", title = featureID, size = 0.5, BPPARAM = bpparam())
}
\arguments{
\item{object}{TargetExperiment object.}

\item{featureID}{Character indicating the ID of the feature.}

\item{SNPs}{Logical flag indicating if SNPs should be plotted.}

\item{xlab}{Character containing the axis x label.}

\item{title}{Character containing the plot title.}

\item{size}{Numeric indicating the size of line plots.}

\item{BPPARAM}{An optional BiocParallelParam instance defining the parallel
back-end to be used during evaluation.}
}
\value{
ggplot2 graphics.
}
\description{
\code{plotFeature} plots the achieved performance for each feature/gene.
The resulting graph shows one bar per each feature/gene with heights
according to its attribute value. If complete is set as TRUE two bar plots
(feature and gene level) will be stored in the resulting ggplot object.


}
\note{
see full example in \code{\link{TargetExperiment-class}}
}
\examples{
if(interactive()){
## loading TargetExperiment object
data(ampliPanel, package="TarSeqQC")
## Defining bam file, bed file and fasta file names and paths
setBamFile(ampliPanel)<-system.file("extdata", "mybam.bam",
    package="TarSeqQC", mustWork=TRUE)
setFastaFile(ampliPanel)<-system.file("extdata", "myfasta.fa",
    package="TarSeqQC", mustWork=TRUE)

# Exploring the read count profile for a particular amplicon
g<-plotFeature(ampliPanel, featureID="AMPL20")
g
}
}
\author{
Gabriela A. Merino \email{gmerino@bdmg.com.ar}, Cristobal Fresno
\email{cfresno@bdmg.com.ar} and Elmer A. Fernandez
\email{efernandez@bdmg.com.ar}
}
\seealso{
\code{\link{plotRegion}}
}

