% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetExperiment-plotGeneAttrPerFeat.R
\docType{methods}
\name{plotGeneAttrPerFeat}
\alias{plotGeneAttrPerFeat}
\alias{plotGeneAttrPerFeat,TargetExperiment-method}
\alias{plotGeneAttrPerFeat-methods}
\title{Plot the attribute value for all the features of a selected gene.}
\usage{
plotGeneAttrPerFeat(object, geneID)

\S4method{plotGeneAttrPerFeat}{TargetExperiment}(object, geneID)
}
\arguments{
\item{object}{TargetExperiment object.}

\item{geneID}{Character indicating the ID of the selected gene.}
}
\value{
ggplot2 graphics.
}
\description{
\code{plotGeneAttrPerFeat} plots the achieved performance for each feature
for a particular gene. The resulting graph shows one bar per each gene feature
with heights according to its attribute value.
}
\note{
see full example in \code{\link{TargetExperiment-class}}
}
\examples{
## Loading the TargetExperiment object
data(ampliPanel, package="TarSeqQC")

# Exploring amplicon attribute values for a particular gene
g<-plotGeneAttrPerFeat(ampliPanel, geneID="gene4")
# Adjust text size
g<-g+theme(title=element_text(size=16), axis.title=element_text(size=16),
legend.text=element_text(size=14))
if(interactive()){
g
}
}
\author{
Gabriela A. Merino \email{gmerino@bdmg.com.ar}, Cristobal Fresno
\email{cfresno@bdmg.com.ar} and Elmer A. Fernandez 
\email{efernandez@bdmg.com.ar}
}
\seealso{
\code{\link{plotAttrExpl}}
}

