% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetExperiment-plotMetaDataExpl.R
\docType{methods}
\name{plotMetaDataExpl}
\alias{plotMetaDataExpl}
\alias{plotMetaDataExpl,TargetExperiment-method}
\alias{plotMetaDataExpl-methods}
\title{Graphical exploration of a specific metadata column.}
\usage{
plotMetaDataExpl(object, name = c("length", "gc", "pool"), log = FALSE,
    join = TRUE, absolute = FALSE, color = "blue")

\S4method{plotMetaDataExpl}{TargetExperiment}(object, name = c("length", "gc",
    "pool"), log = FALSE, join = TRUE, absolute = FALSE, color = "blue")
}
\arguments{
\item{object}{TargetExperiment class object.}

\item{name}{a character indicating the metadata column name that should
be analyzed.}

\item{log}{Logical indicating if the numerical metadata column should be 
considered in log10 scale.}

\item{join}{Logical only for numerical variables. It indicates if boxplot and
density function should be plotted together using the ggplot2 geom_violin 
method.}

\item{absolute}{Logical indicating if the frequencies of the selected 
categorical metadata column should be in absolute scale. If absolute is FALSE
the frequencies are in relative percentages.}

\item{color}{A character indicating a valid name color.}
}
\value{
ggplot2 graphics.
}
\description{
\code{plotMetaDataExpl} plots density and box-plot of a specific metadata 
column. If the characteristic is nonnumerical, then  a frequency plot is 
built.
}
\note{
see full example in \code{\link{TargetExperiment-class}}
}
\examples{
## Loading the TargetExperiment object
data(ampliPanel, package="TarSeqQC")

# Attribute boxplot and density plot exploration
g<-plotMetaDataExpl(ampliPanel, name="length")
if(interactive())
{
# x11(type="cairo")
g
}
# Explore amount of amplicons per gene
g<-plotMetaDataExpl(ampliPanel, name="gene", absolute=TRUE)
if(interactive())
{
# x11(type="cairo")
g
}
}
\author{
Gabriela A. Merino \email{gmerino@bdmg.com.ar}, Cristobal Fresno
\email{cfresno@bdmg.com.ar}, Yanina Murua \email{ymurua@leloir.org.ar},
Andrea S. Llera \email{allera@leloir.org.ar} and Elmer A. Fernandez 
\email{efernandez@bdmg.com.ar}
}
\seealso{
\code{\link{plot}}, \code{\link{plotFeatPerform}}
}

