% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TargetExperiment-print.R
\docType{methods}
\name{print}
\alias{print}
\alias{print,TargetExperiment-method}
\title{Print a TargetExperiment object.}
\usage{
\S4method{print}{TargetExperiment}(x, ...)
}
\arguments{
\item{x}{TargetExperiment class object.}

\item{...}{Included for generic print compatibility.}
}
\value{
console output of the object.
}
\description{
Generic \code{print} method for TargetExperiment class and descendants.
}
\note{
see full example in \code{\link{TargetExperiment-class}}
}
\examples{
## Loading the TargetExperiment object
data(ampliPanel,package="TarSeqQC")
print(ampliPanel)
}
\author{
Gabriela Merino \email{gmerino@bdmg.com.ar}, Cristobal Fresno
\email{cfresno@bdmg.com.ar} and Elmer A. Fernandez
\email{efernandez@bdmg.com.ar}
}

