% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetExperiment-readFrequencies.R
\docType{methods}
\name{readFrequencies}
\alias{readFrequencies}
\alias{readFrequencies,TargetExperiment-method}
\alias{readFrequencies-methods}
\title{Function to explore read frequencies in targeted regions and out 
targeted regions.}
\usage{
readFrequencies(object, BPPARAM = bpparam())

\S4method{readFrequencies}{TargetExperiment}(object, BPPARAM = bpparam())
}
\arguments{
\item{object}{TargetExperiment class object.}

\item{BPPARAM}{An optional BiocParallelParam instance defining the parallel
back-end to be used during evaluation.}
}
\value{
data.frame object.
}
\description{
\code{readFrequencies} builds a data frame containing the read frequencies 
falling in targeted regions and out of those, separated by chromosome.
}
\note{
see full example in \code{\link{TargetExperiment-class}}
}
\examples{
## loading TargetExperiment object
data(ampliPanel, package="TarSeqQC")
## Defining bam file, bed file and fasta file names and paths
setBamFile(ampliPanel)<-system.file("extdata", "mybam.bam", 
    package="TarSeqQC", mustWork=TRUE)
setFastaFile(ampliPanel)<-system.file("extdata", "myfasta.fa", 
    package="TarSeqQC", mustWork=TRUE)

myReadPercentages<-readFrequencies(ampliPanel)
}
\author{
Gabriela A. Merino \email{gmerino@bdmg.com.ar}, Cristobal Fresno
\email{cfresno@bdmg.com.ar}, Yanina Murua \email{ymurua@leloir.org.ar},
Andrea S. Llera \email{allera@leloir.org.ar} and Elmer A. Fernandez 
\email{efernandez@bdmg.com.ar}
}

