% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TargetExperiment-setters.R
\docType{methods}
\name{setFeature<-}
\alias{setAttribute<-}
\alias{setAttribute<-,TargetExperiment,character-method}
\alias{setAttribute<--methods}
\alias{setBamFile<-}
\alias{setBamFile<-,TargetExperiment,character-method}
\alias{setBedFile<-}
\alias{setBedFile<-,TargetExperiment,character-method}
\alias{setBedFile<--methods}
\alias{setFastaFile<-}
\alias{setFastaFile<-,TargetExperiment,character-method}
\alias{setFastaFile<--methods}
\alias{setFeature<-}
\alias{setFeature<-,TargetExperiment,character-method}
\alias{setFeature<--methods}
\alias{setFeaturePanel<-}
\alias{setFeaturePanel<-,TargetExperiment,GRanges-method}
\alias{setFeaturePanel<--methods}
\alias{setGenePanel<-}
\alias{setGenePanel<-,TargetExperiment,GRanges-method}
\alias{setGenePanel<--methods}
\alias{setPileupP<-}
\alias{setPileupP<-,TargetExperiment,PileupParam-method}
\alias{setPileupP<--methods}
\alias{setScanBamP<-}
\alias{setScanBamP<-,TargetExperiment,ScanBamParam-method}
\alias{setScanBamP<--methods}
\alias{setsetBamFile<-}
\alias{setsetBamFile<--methods}
\title{Setters for the TargetExperiment slots}
\usage{
setFeature(object) <- value

\S4method{setFeature}{TargetExperiment,character}(object) <- value

setAttribute(object) <- value

\S4method{setAttribute}{TargetExperiment,character}(object) <- value

setScanBamP(object) <- value

\S4method{setScanBamP}{TargetExperiment,ScanBamParam}(object) <- value

setPileupP(object) <- value

\S4method{setPileupP}{TargetExperiment,PileupParam}(object) <- value

setFeaturePanel(object) <- value

\S4method{setFeaturePanel}{TargetExperiment,GRanges}(object) <- value

setGenePanel(object) <- value

\S4method{setGenePanel}{TargetExperiment,GRanges}(object) <- value

setBedFile(object) <- value

\S4method{setBedFile}{TargetExperiment,character}(object) <- value

setBamFile(object) <- value

\S4method{setBamFile}{TargetExperiment,character}(object) <- value

setFastaFile(object) <- value

\S4method{setFastaFile}{TargetExperiment,character}(object) <- value
}
\arguments{
\item{object}{TargetExperiment class object.}

\item{value}{value to set the slot.}
}
\value{
a TargetExperiment object
}
\description{
Set TargetExperiment slots, according to the given function call.
}
\note{
see full example in \code{\link{TargetExperiment-class}}
}
\examples{
## loading TargetExperiment object
if (interactive()){
    data(ampliPanel, package="TarSeqQC")
    ## Defining bam file, bed file and fasta file names and paths
    setBamFile(ampliPanel)<-system.file("extdata", "mybam.bam",
        package="TarSeqQC", mustWork=TRUE)
    setBedFile(ampliPanel)<-system.file("extdata", "mybed.bed",
        package="TarSeqQC", mustWork=TRUE)
    setFastaFile(ampliPanel)<-system.file("extdata", "myfasta.fa",
        package="TarSeqQC", mustWork=TRUE)
    ## Set feature slot value
    setFeature(ampliPanel)<-"amplicon"
    ## Set attribute slot value
    setAttribute(ampliPanel)<-"coverage"
    ## Set scanBamP slot value
    setScanBamP(ampliPanel)<-ScanBamParam()
    ## Set pileupP slot value
    setPileupP(ampliPanel)<-PileupParam()
    ## Set featurePanel slot value
    setFeaturePanel(ampliPanel)<-buildFeaturePanel(ampliPanel)
    ## Set genePanel slot value
    setGenePanel(ampliPanel)<-summarizePanel(ampliPanel)
    ## Set bedFile slot value
    setBedFile(ampliPanel)<-system.file("extdata", "mybed.bed",
        package="TarSeqQC", mustWork=TRUE)
    ## Set bamFile slot value
    setBamFile(ampliPanel)<-system.file("extdata", "mybam.bam",
        package="TarSeqQC", mustWork=TRUE)
    ## Set fastaFile slot value
    setFastaFile(ampliPanel)<-system.file("extdata", "myfasta.fa",
        package="TarSeqQC", mustWork=TRUE)
}
}
\author{
Gabriela A. Merino \email{gmerino@bdmg.com.ar} Cristobal Fresno
\email{cfresno@bdmg.com.ar} and Elmer A. Fernandez
\email{efernandez@bdmg.com.ar}
}

