% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetExperiment-show.R, 
% R/TargetExperimentList-show.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,TargetExperiment-method}
\alias{show,TargetExperimentList-method}
\title{Show method for the TargetExperiment and TargetExperimentList classes.}
\usage{
\S4method{show}{TargetExperiment}(object)

\S4method{show}{TargetExperimentList}(object)
}
\arguments{
\item{object}{TargetExperiment/TargetExperimentList class object}
}
\value{
console output of the object
}
\description{
\code{show} a TargetExperiment/TargetExperimentList object
}
\details{
Generic show method for TargetExperiment and TargetExperimentList classes 
output visualization.
}
\note{
see full example in \code{\link{TargetExperiment-class}}

see full example in \code{\link{TargetExperimentList-class}}
}
\examples{
## Loading the TargetExperiment object
data(ampliPanel, package="TarSeqQC")
show(ampliPanel)
## Loading the TargetExperimentList object
data(TEList, package="TarSeqQC")
show(TEList)
}
\author{
Gabriela A. Merino \email{gmerino@bdmg.com.ar}, Cristobal Fresno
\email{cfresno@bdmg.com.ar}, Yanina Murua \email{ymurua@leloir.org.ar},
Andrea S. Llera \email{allera@leloir.org.ar} and Elmer A. Fernandez 
\email{efernandez@bdmg.com.ar}
}

