% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TargetExperiment-statistics.R
\docType{methods}
\name{summaryFeatureLev}
\alias{#}
\alias{Object}
\alias{summary}
\alias{summary(ampliPanel)}
\alias{summary,TargetExperiment-method}
\alias{summaryFeatureLev}
\alias{summaryFeatureLev,TargetExperiment-method}
\alias{summaryFeatureLev-methods}
\alias{summaryGeneLev}
\alias{summaryGeneLev,TargetExperiment-method}
\alias{summaryGeneLev-methods}
\alias{summaryIntervals}
\alias{summaryIntervals,TargetExperiment-method}
\alias{summaryIntervals-methods}
\title{TargetExperiment summary.}
\usage{
summaryFeatureLev(object)

\S4method{summaryFeatureLev}{TargetExperiment}(object)

summaryGeneLev(object)

\S4method{summaryGeneLev}{TargetExperiment}(object)

\S4method{summary}{TargetExperiment}(object, ...)

summaryIntervals(object, attributeThres = c(0, 1, 50, 200, 500, Inf))

\S4method{summaryIntervals}{TargetExperiment}(object, attributeThres = c(0, 1,
    50, 200, 500, Inf))
}
\arguments{
\item{object}{TargetExperiment class object.}

\item{...}{required by summary.}

\item{attributeThres}{numeric indicating the intervals extreme values
required by summaryIntervals.}
}
\value{
according to the call one of the following objects can be returned
\item{data.frame}{statistics of the analyzed attribute}
\item{data.frame}{Frequency table of the feature occurrence in the selected
intervals}
}
\description{
Explore the TargetExperiment's attribute values at feature and/or gene level.
}
\note{
see full example in \code{\link{TargetExperiment-class}}
}
\examples{
## Loading the TargetExperiment object
data(ampliPanel, package="TarSeqQC")

# Summary at feature level
summaryFeatureLev(ampliPanel)
# Summary at gene level
summaryGeneLev(ampliPanel)
# Defining the attribute interval extreme values
attributeThres<-c(0,1,50,200,500, Inf)
# Doing a frequency table for the attribute intervals
summaryIntervals(ampliPanel, attributeThres=attributeThres)
}
\author{
Gabriela A. Merino \email{gmerino@bdmg.com.ar}, Cristobal Fresno
\email{cfresno@bdmg.com.ar} and Elmer A. Fernandez
\email{efernandez@bdmg.com.ar}
}

