% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetExperiment-summarizePanel.R
\docType{methods}
\name{summarizePanel}
\alias{summarizePanel}
\alias{summarizePanel,TargetExperiment-method}
\alias{summarizePanel-methods}
\title{Function to summarize a featurePanel slot at a gene level.}
\usage{
summarizePanel(object, BPPARAM = bpparam())

\S4method{summarizePanel}{TargetExperiment}(object, BPPARAM = bpparam())
}
\arguments{
\item{object}{TargetExperiment class object.}

\item{BPPARAM}{An optional BiocParallelParam instance defining the parallel
back-end to be used during evaluation.}
}
\value{
TargetExperiment object
}
\description{
\code{summarizePanel} helps the initialization of a TargetExperiment object.
Is useful to summarize the featurePanel slot at a gene level, building the
genePanel slot.
}
\note{
see full example in \code{\link{TargetExperiment-class}}
}
\examples{
## Loading the TargetExperiment object
data(ampliPanel, package="TarSeqQC")

mySummarizedPanel<-summarizePanel(ampliPanel)
}
\author{
Gabriela A. Merino \email{gmerino@bdmg.com.ar}, Cristobal Fresno
\email{cfresno@bdmg.com.ar}, Yanina Murua \email{ymurua@leloir.org.ar},
Andrea S. Llera \email{allera@leloir.org.ar} and Elmer A. Fernandez 
\email{efernandez@bdmg.com.ar}
}
\seealso{
\code{\link{TargetExperiment}},\code{\link{buildFeaturePanel}}
}

