% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetExperimentList-constructor.R
\docType{methods}
\name{TargetExperimentList}
\alias{TargetExperimentList}
\alias{TargetExperimentList-methods}
\title{TargetExperimentList constructor}
\usage{
TargetExperimentList(TEList, feature = NULL, attribute = "coverage")
}
\arguments{
\item{TEList}{List containing all the TargetExperiment objects corresponding
to the experiments that will be compared.}

\item{feature}{Character indicating the name of the feature that will
be explored (e.g 'amplicon', 'transcript', 'gene').}

\item{attribute}{Character indicating the name of the attribute that will
be explored. Should be 'coverage' or 'medianCounts'.}
}
\value{
TargetExperimentList object.
}
\description{
\code{TargetExperimentList} creates a TargetExperimentList object containing
a set of targeted sequencing experiment results, all those, carried out using
the same bed file. Feature parameter specifies what represent each panel 
element (bed file row). Attribute parameter indicates which attribute would 
be analyzed, 'coverage' or 'medianCounts' and should be specified in order
to indicate which coverage or medianCounts should be conserved.
}
\note{
see full example in \code{\link{TargetExperimentList-class}}
}
\examples{
# Defining the set of TargetExperiment objects
data(ampliPanel, package="TarSeqQC")
data(ampliPanel2, package="TarSeqQC")
ampliList<-list(ampliPanel, ampliPanel2)
# Defining feature parameter
feature<-"amplicon"
# Defining attribute parameter
attribute<-"coverage"
##Calling the constructor
object<-TargetExperimentList(TEList=ampliList, attribute=attribute,
feature=feature)

}
\author{
Gabriela A. Merino \email{gmerino@bdmg.com.ar}, Cristobal Fresno
\email{cfresno@bdmg.com.ar} and Elmer A. Fernandez 
\email{efernandez@bdmg.com.ar}
}
\seealso{
\code{\link{TargetExperimentList-class}}1

Other TargetExperimentList: \code{\link{TargetExperimentList-class}},
    \code{\link{initialize,TargetExperimentList-method}},
    \code{\link{object}}
}

