% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetExperimentList-initialize.R
\docType{methods}
\name{initialize,TargetExperimentList-method}
\alias{TargetExperimentList-method}
\alias{initialize,TargetExperimentList-method}
\title{TargetExperimentList object constructor.}
\usage{
\S4method{initialize}{TargetExperimentList}(.Object, TEList, feature = NULL,
    attribute = "coverage")
}
\arguments{
\item{.Object}{TargetExperimentList class.}

\item{TEList}{List containing all the TargetExperiment objects corresponding
    to the experiments that will be compared.}

\item{feature}{Character indicating the name of the feature that will
    be explored (e.g 'amplicon', 'exon', 'gene').}

\item{attribute}{Character indicating the name of the attribute that will
    be explored. Should be 'coverage' or 'medianCounts'.}
}
\value{
TargetExperimentList object.
}
\description{
\code{initialize} creates the TargetExperimentList object containing the 
    experiment results of several targeted sequencing experiments carried out
    using a unique bed file.
}
\note{
see full example in \code{\link{TargetExperimentList-class}}
}
\examples{
# Defining the set of TargetExperiment objects
data(ampliPanel, package="TarSeqQC")
data(ampliPanel2, package="TarSeqQC")
ampliList<-list(ampliPanel, ampliPanel2)
# Defining feature parameter
feature<-"amplicon"
# Defining attribute parameter
attribute<-"coverage"
##Calling the constructor
object<-TargetExperimentList(TEList=ampliList, attribute=attribute,
    feature=feature)

}
\author{
Gabriela A. Merino \email{gmerino@bdmg.com.ar}, Cristobal Fresno
\email{cfresno@bdmg.com.ar}, Yanina Murua \email{ymurua@leloir.org.ar},
Andrea S. Llera \email{allera@leloir.org.ar} and Elmer A. Fernandez 
\email{efernandez@bdmg.com.ar}
}
\seealso{
\code{\link{TargetExperimentList}}

Other TargetExperimentList: \code{\link{TargetExperimentList-class}},
    \code{\link{TargetExperimentList}}, \code{\link{object}}
}

