% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetExperimentList-plotGlobalAttrExpl.R
\docType{methods}
\name{plotGlobalAttrExpl}
\alias{plotGlobalAttrExpl}
\alias{plotGlobalAttrExpl,TargetExperimentList-method}
\alias{plotGlobalAttrExpl-methods}
\title{Plot attribute exploration of a TargetExperimentList object.}
\usage{
plotGlobalAttrExpl(object, attributeThres = c(0, 1, 50, 200, 500, Inf),
    dens = FALSE, log = FALSE, pool = FALSE, featureLabs = FALSE,
    medianMarg = NULL)

\S4method{plotGlobalAttrExpl}{TargetExperimentList}(object,
    attributeThres = c(0, 1, 50, 200, 500, Inf), dens = FALSE, log = FALSE,
    pool = FALSE, featureLabs = FALSE, medianMarg = NULL)
}
\arguments{
\item{object}{TargetExperimentList class object.}

\item{attributeThres}{Numeric indicating the attribute interval extreme 
values.}

\item{dens}{Logical indicating if boxplot and density function should be 
    plotted together using the ggplot2 geom_violin method or only the boxplot 
    (dens=FALSE) should be displayed.}

\item{log}{Logical indicating if the attribute should be considered in 
    log10 scale.}

\item{pool}{Logical indicating if the plots should be performed for each
    pool separately}

\item{featureLabs}{logical indicating if feature names should be plotted}

\item{medianMarg}{numeric indicating the percentage of the median attribute 
    value to be plotted as lines. If it is NULL no line will be displayed}
}
\value{
ggplot2 graphics.
}
\description{
\code{plotGlobalAttrExpl} displays box-plot of the analyzed achieved 
attribute values along all samples and at a feature level. This graphic
could include density plot together the corresponding  box-plot using the 
ggplot2 geom_violin method.
}
\note{
see full example in \code{\link{TargetExperimentList-class}}
}
\examples{
## Loading the TargetExperimentList object
data(TEList, package="TarSeqQC")

# Attribute boxplot and density plot exploration
g<-plotGlobalAttrExpl(TEList,log=FALSE)
# x11(type="cairo")
if(interactive()){
g
}
}
\author{
Gabriela A. Merino \email{gmerino@bdmg.com.ar}, Cristobal Fresno
\email{cfresno@bdmg.com.ar} and Elmer A. Fernandez 
\email{efernandez@bdmg.com.ar}
}
\seealso{
\code{\link{plot}}
}

