% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetExperimentList-plotPoolPerformance.R
\docType{methods}
\name{plotPoolPerformance}
\alias{plotPoolPerformance}
\alias{plotPoolPerformance,TargetExperimentList-method}
\alias{plotPoolPerformance-methods}
\title{Plot pool performance of a TargetExperimentList object.}
\usage{
plotPoolPerformance(object, dens = FALSE, join = FALSE, log = TRUE,
    attributeThres = NULL)

\S4method{plotPoolPerformance}{TargetExperimentList}(object, dens = FALSE,
    join = FALSE, log = TRUE, attributeThres = NULL)
}
\arguments{
\item{object}{TargetExperimentList class object.}

\item{dens}{Logical indicating if density plot should be included}

\item{join}{Logical indicating if boxplot and density function should be 
plotted together using the ggplot2 geom_violin method. For it uses, dens 
should be TRUE.}

\item{log}{Logical indicating if the attribute should be considered in 
log10 scale.}

\item{attributeThres}{Numeric indicating the attribute interval extreme 
values. It is not a mandatory parameter but if it is specified,then the 
plots will be colored according to the interval in which falls the attribute
median values.}
}
\value{
ggplot2 graphics.
}
\description{
\code{plotPoolPerformance} plots density and/or box-plot of the analyzed
attribute achieved in each PCR pool. These graphics could be displayed 
together using the ggplot2 geom_violin method.
}
\examples{
## Loading the TargetExperimentList object
data(TEList, package="TarSeqQC")
# Attribute boxplot and density plot exploration
g<-plotPoolPerformance(TEList,log=FALSE)
# x11(type="cairo")
if(interactive()){
g
}
}

