% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TarSeqQC-package.R
\docType{package}
\name{TarSeqQC-package}
\alias{TarSeqQC-package}
\title{TarSeqQC: Targeted Sequencing Experiment Quality Control R package}
\description{
The package models targeted sequencing experiment output using previous 
packages. This package includes the new following features: 
\enumerate{
    \item Panel model:
    \itemize{
        \item  Model customizable feature panels.
        \item  Evaluation of the sequencing run performance at median or 
                coverage level for each feature.   
        \item  Exploration of sequenced features. 
    }
    \item Quality Control of the sequencing run:
    \itemize{
        \item General overview of the run performance.
        \item Statistical indicators at median or coverage level.
        \item Xlsx report.
    }
    \item Customizable scan bam file parameters.
    \item Customizable pileup build parameters.
    \item Incorporation of fasta sequence.
    \item Fast exploration of read profile for particular features or genomic
            regions, coloring SNPs occurrences.
}
}
\author{
Gabriela A. Merino \email{gmerino@bdmg.com.ar}, Cristobal Fresno
\email{cfresno@bdmg.com.ar}, Yanina Murua \email{ymurua@leloir.org.ar},
Andrea S. Llera \email{allera@leloir.org.ar} and Elmer A. Fernandez 
\email{efernandez@bdmg.com.ar}
}

