% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetExperiment-ggplotColours.R, 
% R/TargetExperiment-addStatSummSheet.R, R/TargetExperiment-buildReport.R
\docType{methods}
\name{buildReport}
\alias{buildReport}
\alias{buildReport,TargetExperiment-method}
\alias{buildReport-methods}
\alias{ggplotColours}
\alias{ggplotColours,TargetExperiment-method}
\alias{ggplotColours,TargetExperimentList-method}
\alias{ggplotColours-methods}
\alias{addStatSummSheet}
\alias{addStatSummSheet,TargetExperiment-method}
\alias{addStatSummSheet-methods}
\title{TargetExperiment auxiliar function.}
\usage{
buildReport(object, attributeThres = c(0, 1, 50, 200, 500, Inf),
    imageFile = NULL, file= "Results.xlsx")

\S4method{buildReport}{TargetExperiment}(object, attributeThres = c(0, 1, 50,
    200, 500, Inf), imageFile = NULL, file = "Results.xlsx")

ggplotColours(object, n)

\S4method{ggplotColours}{TargetExperiment}(object, n)

\S4method{ggplotColours}{TargetExperimentList}(object, n)

addStatSummSheet(object, wb, attributeThres = c(0, 1, 50, 200, 500, Inf),
    imageFile)

\S4method{addStatSummSheet}{TargetExperiment}(object, wb,
    attributeThres = c(0, 1, 50, 200, 500, Inf), imageFile)
}
\arguments{
\item{object}{TargetExperiment class object.}

\item{attributeThres}{Numeric indicating the intervals extreme values.}

\item{imageFile}{Character indicating the name of the file that contains
the plot that could be insert in the report.}

\item{file}{Character indicating the name of the report.}

\item{n}{amount of colors.}

\item{wb}{A workbook object that will contain the report.}
}
\value{
Workbook object.

colours.
}
\description{

\code{buildReport} builds an excel file containing some statistical results.
These are computed to the selected attribute (e.g. 'coverage') along features
(e.g. 'amplicon') and genes. If 'imageFile' is null, the graph generated
calling the generic plot function will be used.

\code{ggplotColours} is a function to know what color is used when  
ggplot is called.

\code{addStatSummSheet} adds the statistics summary sheet to the workbook
that contains the Target Experiment Report.
}
\note{
see full example in \code{\link{TargetExperiment-class}}
}
\examples{
## Loading the TargetExperiment object
data(ampliPanel,package="TarSeqQC")
# definition of the interval extreme values
attributeThres<-c(0,1,50,200,500, Inf)

## Building the XLSX report
imageFile<-system.file("extdata", "plot.pdf", package="TarSeqQC",
mustWork=TRUE)
buildReport(ampliPanel, attributeThres=attributeThres, imageFile=imageFile,
file="results.xlsx")

## Loading the TargetExperimentList object
data(ampliPanel, package="TarSeqQC")
colors<-ggplotColours(ampliPanel, n=5)
## Loading the TargetExperimentList object
data(TEList, package="TarSeqQC")
colors<-ggplotColours(TEList, n=5)

}
\author{
Gabriela A. Merino \email{gmerino@bdmg.com.ar}, Cristobal Fresno
\email{cfresno@bdmg.com.ar}, Yanina Murua \email{ymurua@leloir.org.ar},
Andrea S. Llera \email{allera@leloir.org.ar} and Elmer A. Fernandez 
\email{efernandez@bdmg.com.ar}
}
\seealso{
\code{\link{TargetExperiment-class}}
}

