% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetExperiment-getters.R,
% R/TargetExperimentList-getters.R
\docType{methods}
\name{getBedFile}
\alias{getAttribute}
\alias{getAttribute,TargetExperiment-method}
\alias{getAttribute,TargetExperimentList-method}
\alias{getAttribute-methods}
\alias{getBamFile}
\alias{getBamFile,TargetExperiment-method}
\alias{getBamFile-methods}
\alias{getBedFile}
\alias{getBedFile,TargetExperiment-method}
\alias{getBedFile,TargetExperimentList-method}
\alias{getBedFile-methods}
\alias{getFastaFile}
\alias{getFastaFile,TargetExperiment-method}
\alias{getFastaFile-methods}
\alias{getFeature}
\alias{getFeature,TargetExperiment-method}
\alias{getFeature,TargetExperimentList-method}
\alias{getFeature-methods}
\alias{getFeaturePanel}
\alias{getFeaturePanel,TargetExperiment-method}
\alias{getFeaturePanel-methods}
\alias{getGenePanel}
\alias{getGenePanel,TargetExperiment-method}
\alias{getGenePanel-methods}
\alias{getLowCtsFeatures}
\alias{getLowCtsFeatures,TargetExperiment-method}
\alias{getLowCtsFeatures,TargetExperimentList-method}
\alias{getLowCtsFeatures-methods}
\alias{getPanels}
\alias{getPanels,TargetExperimentList-method}
\alias{getPileupP}
\alias{getPileupP,TargetExperiment-method}
\alias{getPileupP-methods}
\alias{getRegion}
\alias{getRegion,TargetExperiment-method}
\alias{getOverlappedRegions}
\alias{getOverlappedRegions,TargetExperiment-method}
\alias{getRegion,TargetExperimentList-method}
\alias{getRegion-methods}
\alias{getScanBamP}
\alias{getScanBamP,TargetExperiment-method}
\alias{getScanBamP-methods}
\title{Getters for TargetExperiment and TargetExperimentList objects.}
\usage{
getBedFile(object)

\S4method{getBedFile}{TargetExperiment}(object)

getBamFile(object)

\S4method{getBamFile}{TargetExperiment}(object)

getFastaFile(object)

\S4method{getFastaFile}{TargetExperiment}(object)

getFeaturePanel(object)

\S4method{getFeaturePanel}{TargetExperiment}(object)

getGenePanel(object)

\S4method{getGenePanel}{TargetExperiment}(object)

getFeature(object)

\S4method{getFeature}{TargetExperiment}(object)

getAttribute(object)

\S4method{getAttribute}{TargetExperiment}(object)

getScanBamP(object)

\S4method{getScanBamP}{TargetExperiment}(object)

getPileupP(object)

\S4method{getPileupP}{TargetExperiment}(object)

getRegion(object, level, ID, collapse = TRUE)

\S4method{getRegion}{TargetExperiment}(object, level, ID, collapse = TRUE)

getLowCtsFeatures(object, level, threshold = 50)

\S4method{getLowCtsFeatures}{TargetExperiment}(object, level, threshold = 50)

getOverlappedRegions(object, collapse = FALSE)

\S4method{getOverlappedRegions}{TargetExperiment}(object, collapse = FALSE)

\S4method{getBedFile}{TargetExperimentList}(object)

getPanels(object)

\S4method{getPanels}{TargetExperimentList}(object)

\S4method{getFeature}{TargetExperimentList}(object)

\S4method{getAttribute}{TargetExperimentList}(object)

\S4method{getRegion}{TargetExperimentList}(object, level, ID, collapse = TRUE)

\S4method{getLowCtsFeatures}{TargetExperimentList}(object, level,
    threshold = 50)
}
\arguments{
\item{object}{TargetExperiment/TargetExperimentList class object.}

\item{level}{Character indicating 'gene' or 'feature'. Useful to getRegion
function}

\item{ID}{Character indicating the feature name that getRegion should be 
found.}

\item{collapse}{Logical. Should the region be collapsed?.}

\item{threshold}{Numeric what should be the minimum attribute value?.}
}
\value{
according to the call one of the following objects can be returned
\item{GRanges}{bed file of the experiment}
\item{BamFile}{reference to the BAM file}
\item{FaFile}{reference to the fasta file}
\item{GRanges}{feature panel with statistical information}
\item{GRanges}{summarized version of the feature panel at gene level}
\item{character}{name of the explored features (e.g 'amplicon', 'exon')}
\item{character}{name of the analyzed attribute ('coverage' or 
'medianCounts')}
\item{ScanBamParam}{parameters for the scan of the BAM file}
\item{PileupParam}{parameters for the pileup building}
\item{data.frame}{regions or low counts features}
\item{data.frame}{regions definition for overlapped features}
\item{GRanges}{feature panels with statistical information}
}
\description{
Obtain information about TargetExperiment and TargetExperimentList slots, 
according to the given function call.
}
\note{
see full example in \code{\link{TargetExperiment-class}}
}
\examples{
## Loading the TargetExperiment object
data(ampliPanel,package="TarSeqQC")
## Get the bedFile slot
getBedFile(ampliPanel)
## Get the bamFile slot
getBamFile(ampliPanel)
## Get the fastaFile slot
getFastaFile(ampliPanel)
## Get the feateurePanel slot
getFeaturePanel(ampliPanel)
## Get the genePanel slot
getGenePanel(ampliPanel)
## Get the Feature slot
getFeature(ampliPanel)
## Get the attribute slot
getAttribute(ampliPanel)
## Get the scanBamP slot
getScanBamP(ampliPanel)
## Get the pileupP slot
getPileupP(ampliPanel)
## Get the region related to a feature or a gene
getRegion(ampliPanel, level="gene", ID="gene7", collapse=FALSE)
## Get the low counts features
getLowCtsFeatures(ampliPanel, level="feature")
## Get the regions of overlapped features
getOverlappedRegions(ampliPanel,collapse=FALSE)
## Loading the TargetExperimentList object
data(TEList,package="TarSeqQC")
## Get the bedFile slot
getBedFile(TEList)
## Get the panels slot
getPanels(TEList)
## Get the Feature slot
getFeature(TEList)
## Get the attribute slot
getAttribute(TEList)
## Get the region related to a feature or a gene
getRegion(TEList, level="gene", ID="gene7", collapse=FALSE)

## Get the low counts features
getLowCtsFeatures(TEList, level="feature")
}
\author{
Gabriela A. Merino \email{gmerino@bdmg.com.ar}, Cristobal Fresno
\email{cfresno@bdmg.com.ar}, Yanina Murua \email{ymurua@leloir.org.ar},
Andrea S. Llera \email{allera@leloir.org.ar} and Elmer A. Fernandez 
\email{efernandez@bdmg.com.ar}
}
\seealso{
\code{\link{TargetExperiment-class}}
}

