% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TargetExperiment-plotAttrExpl.R
\docType{methods}
\name{plotAttrExpl}
\alias{plotAttrExpl}
\alias{plotAttrExpl,TargetExperiment-method}
\alias{plotAttrExpl-methods}
\title{Plot attribute exploration of a TargetExperiment object.}
\usage{
plotAttrExpl(object, level = "feature", join = TRUE, log = TRUE,
    color = "blue")

\S4method{plotAttrExpl}{TargetExperiment}(object, level = "feature",
    join = TRUE, log = TRUE, color = "blue")
}
\arguments{
\item{object}{TargetExperiment class object.}

\item{level}{Character 'feature' or 'gene' indicating at which level should
be analyzed the attribute.}

\item{join}{Logical indicating if boxplot and density function should be
plotted together using the ggplot2 geom_violin method.}

\item{log}{Logical indicating if the attribute should be considered in
log10 scale.}

\item{color}{A character indicating a valid name color.}
}
\value{
ggplot2 graphics.
}
\description{
\code{plotAttrExpl} plots density and box-plot of the analyzed attribute at a
feature or gene level. This graphics could plot together using the ggplot2
geom_violin method.


}
\note{
see full example in \code{\link{TargetExperiment-class}}
}
\examples{
## Loading the TargetExperiment object
data(ampliPanel, package="TarSeqQC")

# Attribute boxplot and density plot exploration
g<-plotAttrExpl(ampliPanel,level="feature",join=TRUE, log=FALSE, color="blue")
# x11(type="cairo")
if(interactive()){
g
}
}
\author{
Gabriela A. Merino \email{gmerino@bdmg.com.ar}, Cristobal Fresno
\email{cfresno@bdmg.com.ar} and Elmer A. Fernandez
\email{efernandez@bdmg.com.ar}
}
\seealso{
\code{\link{plot}}, \code{\link{plotFeatPerform}}
}

