% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetExperiment-plotFeature.R
\docType{methods}
\name{plotFeature}
\alias{plotFeature}
\alias{plotFeature,TargetExperiment-method}
\alias{plotFeature-methods}
\title{Plot read profiles for a particular feature.}
\usage{
plotFeature(object, featureID, SNPs = TRUE, minAAF=0.05, minRD=10, xlab = "",
    title = "",  size = 0.5, BPPARAM = bpparam())

\S4method{plotFeature}{TargetExperiment}(object, featureID, SNPs = TRUE,
    minAAF=0.05, minRD=10, xlab = "", title = featureID, size = 0.5, 
    BPPARAM = bpparam())
}
\arguments{
\item{object}{TargetExperiment object.}

\item{featureID}{Character indicating the ID of the feature.}

\item{SNPs}{Logical flag indicating if SNPs should be plotted.}

\item{minAAF}{Numeric indicating the minimum alternative allele proportion 
necessary to call a SNP.}

\item{minRD}{Numeric indicating the minimum read depth of alternative 
alleles necessary to call a SNP.}

\item{xlab}{Character containing the axis x label.}

\item{title}{Character containing the plot title.}

\item{size}{Numeric indicating the size of line plots.}

\item{BPPARAM}{An optional BiocParallelParam instance defining the parallel
back-end to be used during evaluation.}
}
\value{
ggplot2 graphics.
}
\description{
\code{plotFeature} plots the read profiles for a selected feature. The minAAF 
parameter set the minimum proportion value to call an SNP and the minRD the 
minimum read depth. They are combined to  obtain a minimum read count value at
each position used to distinguish between possible SNPs and background noise. 
If SNPs is set as 'TRUE', colored bars will appear indicating the occurrence of
possible SNPs surpassing the minAAF and minRD, at each genomic position.
}
\note{
see full example in \code{\link{TargetExperiment-class}}
}
\examples{
## loading TargetExperiment object
data(ampliPanel, package="TarSeqQC")
## Defining bam file, bed file and fasta file names and paths
setBamFile(ampliPanel)<-system.file("extdata", "mybam.bam", 
    package="TarSeqQC", mustWork=TRUE)
setFastaFile(ampliPanel)<-system.file("extdata", "myfasta.fa", 
    package="TarSeqQC", mustWork=TRUE)

# Exploring the read count profile for a particular amplicon
g<-plotFeature(ampliPanel, featureID="AMPL20")
if(interactive()){
g
}
}
\author{
Gabriela A. Merino \email{gmerino@bdmg.com.ar}, Cristobal Fresno
\email{cfresno@bdmg.com.ar}, Yanina Murua \email{ymurua@leloir.org.ar},
Andrea S. Llera \email{allera@leloir.org.ar} and Elmer A. Fernandez 
\email{efernandez@bdmg.com.ar}
}
\seealso{
\code{\link{plotRegion}}
}

