% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetExperiment-plotRegion.R
\docType{methods}
\name{plotRegion}
\alias{plotRegion}
\alias{plotRegion,TargetExperiment-method}
\alias{plotRegion-methods}
\title{Plot read profiles for a particular genomic region.}
\usage{
plotRegion(object, region, seqname, SNPs = TRUE, minAAF=0.05, minRD=10, 
    xlab = "", title = "", size = 0.5, BPPARAM = bpparam())

\S4method{plotRegion}{TargetExperiment}(object, region, seqname, SNPs = TRUE,
    minAAF=0.05, minRD=10, xlab = "", title = "", size = 0.5, 
    BPPARAM = bpparam())
}
\arguments{
\item{object}{TargetExperiment object.}

\item{region}{Numeric of length two indicating the selected genomic region.}

\item{seqname}{Character indicating the chromosome of the genomic region.}

\item{SNPs}{Logical flag indicating if SNPs should be plotted.}

\item{minAAF}{Numeric indicating the minimum alternative allele proportion 
necessary to call a SNP.}

\item{minRD}{Numeric indicating the minimum read depth of alternative 
alleles necessary to call a SNP.}

\item{xlab}{Character containing the axis x label.}

\item{title}{Character containing the plot title.}

\item{size}{Numeric indicating the size of line plots.}

\item{BPPARAM}{An optional BiocParallelParam instance defining the parallel
back-end to be used during evaluation.}
}
\value{
ggplot2 graphics.

include TargetExperiment-FeatPerform.R
}
\description{
\code{plotRegion} plots the read profiles for a selected region. The minAAF 
parameter set the minimum proportion value to call an SNP and the minRD the 
minimum read depth. They are combined to  obtain a minimum read count value at
each position used to distinguish between possible SNPs and background noise. 
If SNPs is set as 'TRUE', colored bars will appear indicating the occurrence of
possible SNPs surpassing the minAAF and minRD, at each genomic position.
}
\note{
see full example in \code{\link{TargetExperiment-class}}
}
\examples{
## loading TargetExperiment object
data(ampliPanel, package="TarSeqQC")
## Defining bam file, bed file and fasta file names and paths
setBamFile(ampliPanel)<-system.file("extdata", "mybam.bam", 
    package="TarSeqQC", mustWork=TRUE)
setFastaFile(ampliPanel)<-system.file("extdata", "myfasta.fa", 
    package="TarSeqQC", mustWork=TRUE)

# getting and exploring a sequenced region of a particular gene
getRegion(ampliPanel, level="gene", ID="gene7", collapse=FALSE)
# plot a particular genomic region
g<-plotRegion(ampliPanel,region=c(4500,6800), seqname="chr10", SNPs=TRUE,  
xlab="", title="gene7 amplicons",size=0.5)
# x11(type="cairo")
if(interactive()){
g
}
}
\author{
Gabriela A. Merino \email{gmerino@bdmg.com.ar}, Cristobal Fresno
\email{cfresno@bdmg.com.ar}, Yanina Murua \email{ymurua@leloir.org.ar},
Andrea S. Llera \email{allera@leloir.org.ar} and Elmer A. Fernandez 
\email{efernandez@bdmg.com.ar}
}

\seealso{
\code{\link{plotFeature}}
}

