% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetExperiment-print.R, 
% R/TargetExperimentList-print.R
\docType{methods}
\name{print}
\alias{print}
\alias{print,TargetExperiment-method}
\alias{print,TargetExperimentList-method}
\title{Print a TargetExperiment/TargetExperimentList object.}
\usage{
\S4method{print}{TargetExperiment}(x, ...)

\S4method{print}{TargetExperimentList}(x, ...)
}
\arguments{
\item{x}{TargetExperiment/TargetExperimentList class object.}

\item{...}{Included for generic print compatibility.}
}
\value{
console output of the object.
}
\description{
Generic \code{print} method for TargetExperiment and TargetExperimentList 
classes and descendants.
}
\note{
see full example in \code{\link{TargetExperiment-class}}

see full example in \code{\link{TargetExperimentList-class}}
}
\examples{
## Loading the TargetExperiment object
data(ampliPanel,package="TarSeqQC")
print(ampliPanel)
## Loading the TargetExperimentList object
data(TEList,package="TarSeqQC")
print(TEList)
}
\author{
Gabriela A. Merino \email{gmerino@bdmg.com.ar}, Cristobal Fresno
\email{cfresno@bdmg.com.ar}, Yanina Murua \email{ymurua@leloir.org.ar},
Andrea S. Llera \email{allera@leloir.org.ar} and Elmer A. Fernandez 
\email{efernandez@bdmg.com.ar}
}

