% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TargetExperiment-show.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,TargetExperiment-method}
\title{Show method for the TargetExperiment class.}
\usage{
\S4method{show}{TargetExperiment}(object)
}
\arguments{
\item{object}{TargetExperiment class object}
}
\value{
console output of the object
}
\description{
\code{show} a TargetExperiment object
}
\details{
Generic show method for TargetExperiment class output visualization.
}
\note{
see full example in \code{\link{TargetExperiment-class}}
}
\examples{
## Loading the TargetExperiment object
data(ampliPanel, package="TarSeqQC")
show(ampliPanel)
}
\author{
Gabriela Merino \email{gmerino@bdmg.com.ar}, Cristobal Fresno
\email{cfresno@bdmg.com.ar} and Elmer A. Fernandez
\email{efernandez@bdmg.com.ar}
}

